crew/requirements.txt
crewai
crewai[tools]
python-dotenv
langchain-groq
langchain-openai
python-dotenv
